feat(crm): el contacto pasa a ser el eje, rubro de alta gama y cotizaciones
El CRM deja de ser B2B. En una boutique de alta gama el cliente es una persona
con nombre y teléfono, no una razón social: quien compra un reloj de doce
millones lo hace a título personal y vuelve tres años después por otro. El
contacto absorbe estado, fuente, dueño y etiquetas; la empresa queda para el
regalo corporativo, que es la excepción.

- Nuevo mock: Belmont Alta Relojería, boutique de relojería fina y alta joyería
  con showroom en Alonso de Córdova. Marcas y clientes inventados; se eligió el
  rubro por su mecánica de venta —ticket alto, consultiva por WhatsApp, cita en
  showroom, coleccionista que vuelve— y no por el producto.
- Módulo Contactos: cartera con etiquetas y búsqueda sin tildes, ficha con
  compras, cotizaciones, oportunidades, preferencias y potencial explicado.
- Cotizaciones de mostrador, portadas de CDC: validación única compartida entre
  crear y editar, precios releídos del catálogo (nunca del formulario),
  descuento por pieza con tope + descuento global sobre el remanente, nombre y
  precio congelados al cotizar, y la marca de "editada tras enviarse".
- El puntaje usa ventana de 24 meses: con 12, un coleccionista que compró hace
  catorce puntuaba igual que alguien que nunca compró, y es justo a quien hay
  que llamar. El factor de potencial mira la pieza más cara comprada, mejor
  predictor que el acumulado.

Bug de rango encontrado al cambiar de rubro: los montos pasaron de millones a
decenas de millones y `sum(monto * probabilidad)` desbordaba el integer de
Postgres (tope 2.147.483.647), tumbando la portada con "integer out of range".
Las agregaciones de dinero pasan a float8 —exacto hasta 2^53 y vuelve como
number—; los contadores siguen en int.

+++ lib/crm/cuentas.ts
@@ -66,8 +66,8 @@ export async function listarCuentas(opciones?: {
       owner: crmUsers.nombre,
       contactos: sql<number>`(select count(*) from crm_contacts c where c.account_id = crm_accounts.id)::int`,
       dealsAbiertos: sql<number>`(select count(*) from crm_deals d where d.account_id = crm_accounts.id and d.etapa not in ('ganado','perdido'))::int`,
-      montoAbierto: sql<number>`coalesce((select sum(d.monto) from crm_deals d where d.account_id = crm_accounts.id and d.etapa not in ('ganado','perdido')),0)::int`,
-      facturado: sql<number>`coalesce((select sum(o.total) from crm_orders o where o.account_id = crm_accounts.id),0)::int`,
+      montoAbierto: sql<number>`coalesce((select sum(d.monto) from crm_deals d where d.account_id = crm_accounts.id and d.etapa not in ('ganado','perdido')),0)::float8`,
+      facturado: sql<number>`coalesce((select sum(o.total) from crm_orders o where o.account_id = crm_accounts.id),0)::float8`,
       ultimaCompra: sql<Date | null>`(select max(o.fecha) from crm_orders o where o.account_id = crm_accounts.id)`,
       ultimaActividad: sql<Date | null>`(select max(a.ocurrido_en) from crm_activities a where a.account_id = crm_accounts.id)`,
     })
